Moved base tag population script to scripts folder
